% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{aicbic.IPE}
\alias{aicbic.IPE}
\title{Dimension Selection of IPE by AIC&BIC.}
\usage{
aicbic.IPE(X, Y, NonGrass = F, trace = 0)
}
\arguments{
\item{NonGrass}{Whether to implement further optimization in each dimension?}
}
\value{
A list.
\item{d.aic}
\item{d.bic}
\item{loglik.seq}
\item{aic.seq}
\item{bic.seq}
}
\description{
Dimension Selection of IPE by AIC&BIC.
}
