% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alternating Algorithm.R
\name{alt_alg.p}
\alias{alt_alg.p}
\title{Revised version of \code{\link{alt_alg}}.}
\usage{
alt_alg.p(
  u1,
  u2,
  r,
  M,
  invMU,
  ca,
  ce,
  cb1,
  GEidx.g1,
  GEidx.g12,
  GEidx.cb1,
  obj1
)
}
\description{
Additionally print the value of the objective function after each round of optimization.
}
