% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alternating Algorithm.R
\name{alt_alg}
\alias{alt_alg}
\title{Alternating Algorithm.}
\usage{
alt_alg(u1, u2, r, M, invMU, ca, ce, cb1, GEidx.g1, GEidx.g12, GEidx.cb1, obj1)
}
\arguments{
\item{u1, u2}{u1+u2<r.}
}
\description{
Alternating Algorithm.
}
