% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminaries.R
\name{angle}
\alias{angle}
\title{An "Angle" Between Two Semi-orthogonal Bases}
\usage{
angle(base1, base2)
}
\arguments{
\item{base1, base2}{Two semi-orthogonal matrices.}
}
\value{
The value of the "angle".
}
\description{
A possible problem: when measuring results by data of population version, the
singular value may exceed 1 and thus \code{NaN} is produces by \code{\link{acos}}.
}
