% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{avar_man.IPE}
\alias{avar_man.IPE}
\title{Calculate asymptotic variance of IPE}
\usage{
avar_man.IPE(data.pop.IPE, n = 1000, d, r, p)
}
\description{
See \code{\link{avar_sc.IPE}} for details.
}
