% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{avar_ori.IPE}
\alias{avar_ori.IPE}
\title{Calculate asymptotic variance of IPE}
\usage{
avar_ori.IPE(data.pop.IPE, n = 1000, d, r, p)
}
\arguments{
\item{data.pop.IPE}{In the data, \eqn{d \geq 2} is required, for \code{matrixcalc::K.matrix}
to calculate the communication matrix.}
}
\description{
See \code{\link{avar_man.IPE}} for details. The only difference is that the data, \code{data.pop.IPE}
must have a \eqn{d \geq 2}.
}
