% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{avar_sc.IPE}
\alias{avar_sc.IPE}
\title{Calculate asymptotic variance of IPE for special case}
\usage{
avar_sc.IPE(data.pop.IPE, n = 1000, d, p)
}
\arguments{
\item{data.pop.IPE}{Object of class inheriting from \code{\link{data_gen.IPE}}.
Population version is fine.}

\item{n}{The supposed sample size.}
}
\value{
\item{avarbetaIPE}
\item{avarbetaOLS}
\item{ratio_diag_OLS_IPE}
}
\description{
Calculate asymptotic variance of IPE for special case
}
