% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{boot_dat.inp}
\alias{boot_dat.inp}
\title{Bootstrap Standard Errors for inner PLS}
\usage{
boot_dat.inp(X, Y, d, B)
}
\value{
The bootstrap standard errors.
}
\description{
Basically same to \code{\link{boot_res.inp}}, except that this is by
bootstrapping the original X and Y matrices.
}
