% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{boot_dat}
\alias{boot_dat}
\title{Bootstrap Standard Errors for the inner PLS, IPE, Predictor envelopes and PLS.}
\usage{
boot_dat(X, Y, d, u, u.PLS, B, NonGrass = F, ftol.nonG = 0.1, PRINT = F)
}
\arguments{
\item{d}{Dimension of the inner predictor envelope subspace.}

\item{u}{Dimension of the predictor envelope subspace.}

\item{u.PLS}{Dimension of the SIMPLS algorithm.}

\item{NonGrass}{Whether to use Non-Grassmann technique in \code{IPE} when
estimating the IPE.}

\item{ftol.nonG}{Used in \code{IPE}.}
}
\value{
\item{bootse.inp}
\item{bootpval.inp}
\item{bootse.IPE}
\item{bootpval.IPE}
\item{bootse.predenv}
\item{bootpval.predenv}
\item{bootse.PLS}
\item{bootpval.PLS}
}
\description{
Computes the bootstrap standard errors for each element of the regression
coefficients by bootstrapping the original data matrices X and Y.
}
