% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{boot_res.inp}
\alias{boot_res.inp}
\title{Bootstrap Standard Errors for inner PLS}
\usage{
boot_res.inp(X, Y, d, B)
}
\arguments{
\item{d}{Dimension of the inner envelope subspace. Can be 0? But may need
further debugging.}
}
\value{
The bootstrap standard errors.
}
\description{
Computes the bootstrap standard errors for each element of the regression
coefficients by bootstrapping the residuals.
}
\details{
Learned from \code{\link[Renvlp]{boot.env}}.
}
