% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions in Pkg-Renvlp.R
\name{contr_rd}
\alias{contr_rd}
\title{Contraction Matrix}
\usage{
contr_rd(d)
}
\description{
This function seems to has been deleted from the latest \code{Renvlp} package.
}
\details{
The original function in \code{Renvlp} had the name without "_rd".
}
