% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{d.boot.sim}
\alias{d.boot.sim}
\title{Dimension Selection of inner PLS, IPE, PLS and Predictor Envelope by Bootstrap.}
\usage{
d.boot.sim(
  X,
  Y,
  B = 50,
  thres = 0.9,
  estmrs = c("iPLS", "IPE", "PredEnv", "SIMPLS"),
  trace = 0
)
}
\arguments{
\item{thres}{The threshold to use.}

\item{estmrs}{A character string containing the estimators of which to choose
dimension.}

\item{trace}{If 1, information is printed. If -1, a "-" is printed after the
whole function ends. The -1 part is useful when replicating this function.}
}
\value{
A list. If the method is not performed, then no return values. If no
estimators is specified, then \code{NULL} is returned.
\item{d.boot.IPE}{The selected dimension. If not any of the vector correlations
exceeded the \code{thres}, \code{NA} will be returned}
\item{dis.IPE}{The bootstrap mean of vector correlation value of each dimension.}
\item{d.boot.iPLS}
\item{dis.iPLS}
\item{u.boot.PLS}
\item{dis.PLS}
\item{u.boot.PredEnv}
\item{dis.PredEnv}
}
\description{
Dimension Selection of inner PLS, IPE, PLS and Predictor Envelope by bootstrapping
the data to acquire difference bases. Finally choose a dimension with relatively
the smallest variability.
}
\details{
Implement bootstrap to get the average distance, which is measured by vector
correlation, between each bootstrapped basis and the basis obtained from the
original data. Then select the largest(predictor envelope and PLS) or
smallest(inner PLS and IPE) dimension with corresponding distance over
\code{thres}.
}
