% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data.test}
\alias{data.test}
\title{To see if the data by \code{\link{data_gen.rdenv}} can be estimated well by the \code{\link[Renvlp]{envMU}} method.}
\usage{
data.test(PRINT = F, type = "po", u1, u2, p, r, n, x, dia)
}
\arguments{
\item{PRINT}{If true,the process of choosing starting values will be printed.}
}
\value{
\code{list(d12.sv=d12.sv, a12.sv=a12.sv, d1.sv=d1.sv, a1.sv=a1.sv, d12.clc=d12.clc, a12.clc=a12.clc, d1.clc=d1.clc, a1.clc=a1.clc, noise.beta=noise.beta, noise.M=noise.M)}.
}
\description{
Pu by \code{\link[Renvlp]{ginv}}: \code{tol=.Machine$double.eps^0.8}.
}
