% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.ienv}
\alias{data_gen.ienv}
\title{Data Generation for Inner Envelope Simulation}
\usage{
data_gen.ienv(SAM = F, d, p, r, dia, n, x, upper.eta = 10, VERBOSE = F)
}
\arguments{
\item{d}{Dimension of the inner envelope subspaces. \eqn{d=p} is allowed.}

\item{dia}{Eigenvalues of \eqn{M}, with the first \eqn{d} ones corresponding
to the \eqn{S_1} subspace.}

\item{x}{Used as \code{X<-MASS::mvrnorm(n, rep(0, p), x*diag(p))}, to generate
sample data matrix X.}

\item{VERBOSE}{Whether to print some extra warning information.}
}
\description{
The envelope subspace will always be the full space.
}
\details{
\eqn{\eta_1, \eta_2} by \code{runif(.., 0, upper.eta)}.
}
