% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.ienv2}
\alias{data_gen.ienv2}
\title{Data Generation for Inner Envelope Simulation}
\usage{
data_gen.ienv2(
  SAM = F,
  d,
  p,
  r,
  dia,
  n,
  x,
  B = NULL,
  upper.eta = 10,
  VERBOSE = F
)
}
\arguments{
\item{B}{The \eqn{B} matrix. When \code{d=p} this input is useless.}
}
\description{
Improved version of \code{\link{data_gen.ienv}}. The \eqn{B} matrix can be
specified in the input.
}
