% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.IPE}
\alias{data_gen.IPE}
\title{Data Generation for Inner Predictor Envelope Simulation}
\usage{
data_gen.IPE(
  SAM = F,
  d,
  p,
  r,
  eig.SiX,
  eig.SiYvX,
  n,
  upper.eta = 10,
  VERBOSE = F,
  special_case = F,
  coef.eta2 = 100
)
}
\arguments{
\item{SAM}{Can only generate sample data with normal errors.}

\item{d}{Dimension of the inner envelope subspaces. \eqn{d=r} is allowed. But
\eqn{d>0}a is required.}

\item{eig.SiX}{Eigenvalues of \eqn{\Sigma_X}, with the first \eqn{d} ones corresponding
to the \eqn{S_1} subspace.}

\item{eig.SiYvX}{Eigenvalues of \eqn{\Sigma_{Y|X}}. This matrix is generated
by \code{A diag(eigSiYvX) A^T} with \eqn{A} being random generated orthogonal
matrix. Moreover, this matrix will only be used when generating sample data.}

\item{VERBOSE}{Whether to print some extra warning information.}

\item{special_case}{\eqn{\eta} is generated to satisfy the setting in the special case, but only
suitable for the case when SiYvX has same diagonal elements.. Specifically,
they are generated by 'cutting' an orthogonal matrix.}

\item{coef.eta2}{Coefficient added to \eqn{\eta_2} to make \eqn{\beta} larger
in the \strong{special case}.}
}
\description{
The envelope subspace will always be the full space. Notice that \eqn{\beta}
is the transposed version! That is, the one used in Cook's notation for predictor
envelopes.
}
\details{
If \eqn{d=r}, then \eqn{B,\eta_2} will be \code{NA}.

\eqn{\eta_1, \eta_2} by \code{runif(.., 0, upper.eta)} when \code{special_case=F}.
}
