% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.IPE2}
\alias{data_gen.IPE2}
\title{Data generation for speicial case of IPE. OBSOLETE.}
\usage{
data_gen.IPE2(
  SAM = F,
  d,
  p,
  r,
  eig.SiX,
  eig.SiYvX,
  n,
  coef.eta2 = 100,
  VERBOSE = F
)
}
\description{
Use \code{\link{data_gen.IPE}}.
}
