% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.pred_env}
\alias{data_gen.pred_env}
\title{Data Generation for Predictor Envelope Simulation}
\usage{
data_gen.pred_env(
  SAM = F,
  u,
  r,
  p,
  eig.SiX,
  eig.SiYvX,
  n,
  upper.alpha = 10,
  VERBOSE = F
)
}
\arguments{
\item{u}{Dimension of the envelope subspace.}

\item{eig.SiX}{Eigenvalues of \eqn{\Sigma_X}, with the first \eqn{u} ones corresponding
to the envelope subspace.}

\item{eig.SiYvX}{Eigenvalues of \eqn{\Sigma_{Y|X}}. This matrix is generated
by \code{A diag(eigSiYvX) A^T} with \eqn{A} being random generated orthogonal
matrix.}

\item{VERBOSE}{Whether to print some extra warning information.}
}
\value{
The reuturned \code{M, U} matrix is the one directly derived from the objective
function, on a general envelope sight. By simple envelope algebra it can be shown
that \eqn{E_M(U)} equals the predictor envelope subspace.
}
\description{
Notice that \eqn{\beta} is the transposed version! That is, the one used in
Cook's notation for predictor envelopes.
}
\details{
\eqn{\alpha} by \code{runif(.., 0, upper.eta)}.
}
