% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{data_gen.rdenv}
\alias{data_gen.rdenv}
\title{Data Generation for Russian Doll Envelope Simulation.}
\usage{
data_gen.rdenv(SAM = F, u1, u2, p, r, n, x, dia, upper.eta = 10)
}
\arguments{
\item{SAM}{Whether to use sample version.}

\item{u1, u2}{u1<=p; u1+u2>=p; u2!=1. Because we assume that beta is of full rank.}

\item{upper.eta}{The upper bound for the distribution of eta, which is a
uniform one, with the lower bound being 0.}
}
\description{
Gamma1 corresponds to the first u1 eigenvalues of Sigma, Gamma2 corrsponds to
the next u2 ones, and Gamma3 the rest.
}
