% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation of Simulation.R
\name{data_READ}
\alias{data_READ}
\title{Read Parameters of a IPE Model from a Folder}
\usage{
data_READ(fdnm, work_dir0 = "Y:/Research/Data Generated and Results/IPE")
}
\arguments{
\item{fdnm}{The name of the folder containing the txt files.}

\item{work_dir0}{The working directory where the folder is.}
}
\value{
\item{dat}{A list containing the parameters.}
}
\description{
Reads \code{"Gamma1", "Gamma0", "C", "beta", "eta1", "eta2", "eig.SiX", "eig.SiYvX"}.
Also restores \code{SiX} and \code{SiYvX}
}
