% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation of Simulation.R
\name{data_save}
\alias{data_save}
\title{Save Parameters of a IPE Model as txt Files}
\usage{
data_save(
  dat,
  work_dir0 = "Y:/Research/Data Generated and Results/IPE",
  SC = F
)
}
\arguments{
\item{dat}{The R object containing the parameters.}

\item{work_dir0}{The working directory to write the data on.}

\item{SC}{Whether it is a special case?}
}
\value{
A folder at \code{work_dir0}, the name of which is
\code{"d", d, "r", r, "p", p, "n", n, " ", time_now}.
}
\description{
Saves \code{Gamma1, Gamma0, beta, eta1, eta2, C, eig.SiX, eig.SiYvX} as txt
files. Notice only the eigenvalues--\code{eig.SiX, eig.SiYvX}--are saved.
}
\details{
Notice that \code{setwd} function is used! Be cautious of possible
side effects!
}
