% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions in Pkg-Renvlp.R
\name{envMU2}
\alias{envMU2}
\title{Improved Version of \code{\link[Renvlp]{envMU}}.}
\usage{
envMU2(M, U, u, initial = NULL)
}
\description{
Minor mistakes revised so \code{initial} can be set. And the starting values
used will be returned, for further study, maybe.
}
