% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions in Pkg-Renvlp.R
\name{envMU_print}
\alias{envMU_print}
\title{Improved Version of \code{\link{envMU2}}.}
\usage{
envMU_print(M, U, u, initial = NULL)
}
\description{
Print the values of the calculated objective function, which is used as the
convergence criterion. While temporarily I just deleted the \code{gobj} in
\code{optim}.
}
