% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminaries.R
\name{frbnorm}
\alias{frbnorm}
\title{Compute the Frobenius norm between two bases.}
\usage{
frbnorm(base1, base2)
}
\arguments{
\item{base1, base2}{Two semi-orthogonal basis matrices.}
}
\value{
The Frobenius norm of the difference of their projection matrices.
}
\description{
Input the bases will suffice!
}
