% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{get_GEE}
\alias{get_GEE}
\title{Using GEE to estimate inner envelope subspace.}
\usage{
get_GEE(u, X, Y, num = 200)
}
\arguments{
\item{u}{is the dimension of the inner envelope subspace. u!=0 is required.}
}
\description{
Written by Linquan Ma.
}
\examples{
\dontrun{
#---Usage---
out<-get_GEE(u, dat$X, dat$Y, 50)
out1<-out[-(1:(u*(p-u)))]
out2<-out[1:(u*(p-u))]
Gamma0<-qr.Q(qr(rbind(diag(r-p), matrix(out1, ncol=r-p))))
Gamma1<-qr.Q(qr(Gamma0), complete=T)[, (r-p+1):r]
B<-qr.Q(qr(rbind(diag(u), matrix(out2, ncol=u))))
Gam1<-Gamma1\%*\%B  # This is the Estimator of Gamma1
}
}
