% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv0_logM}
\alias{ienv0_logM}
\title{Inner envelope estimation.}
\usage{
ienv0_logM(M, U, d, initial)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv0}}.
}
\details{
\code{fobj} is in its original form. \eqn{log|M|} is additionally added to
\code{obj1, obj5}.
}
