% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv1}
\alias{ienv1}
\title{Inner envelope estimation.}
\usage{
ienv1(M, U, d, initial)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv0}}.
}
\details{
The inverse in \code{fobj} is improved.
}
