% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv1_logM}
\alias{ienv1_logM}
\title{Inner envelope estimation.}
\usage{
ienv1_logM(M, U, d, initial)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv1}}.
}
\details{
The inverse in \code{fobj} is improved. \eqn{log|M|} is additionally added to
 \code{obj1, obj5}.
}
