% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv1_logM_sym}
\alias{ienv1_logM_sym}
\title{Inner envelope estimation.}
\usage{
ienv1_logM_sym(M, U, d, initial, ftol = 0.001, PRINT = F)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv1}}.
}
\details{
The inverse in \code{fobj} is improved. \eqn{log|M|} is additionally added to
 \code{obj1, obj5}. \code{symmetric=F} is set in \code{eigen}. This has been the
draft for my star function.
}
