% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv1_logM_sym_more}
\alias{ienv1_logM_sym_more}
\title{Inner envelope estimation.}
\usage{
ienv1_logM_sym_more(M, U, d, initial)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv1_logM_sym}}.
}
\details{
Further improve \code{fobj}. Seems to add lots of brackets in fobj? I have
forgotten...
}
