% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienv2_logM}
\alias{ienv2_logM}
\title{Inner envelope estimation.}
\usage{
ienv2_logM(M, U, d, initial)
}
\arguments{
\item{initial}{Starting values for Gamma0!}
}
\value{
\code{Gamma1, Gamma0, B, Omega0}.
}
\description{
Revised version of \code{\link{ienv2}}.
}
\details{
The inverse in \code{fobj} is improved. The symmetry in \code{obj1, obj5}
is improved. And \eqn{log|M|} is additionally added to \code{obj1, obj5}.
}
