% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{ienvMU}
\alias{ienvMU}
\title{Inner envelope estimation. USE ME!}
\usage{
ienvMU(M, U, d, initial, betaOLS = NULL, ftol = 0.01, PRINT = F)
}
\arguments{
\item{initial}{Starting values for \strong{Gamma0} and must be specified.}

\item{ftol}{Used in \code{abs(obj1-obj5)<ftol*abs(obj1)}.}

\item{PRINT}{Print the value of the objective function in each iteration?}
}
\value{
\code{Gamma1, Gamma0, B, Omega0, objfun, betaIE}.
}
\description{
Estimating inner envelope subspace using non-Grassmann technique, which, on
on the same time, relies heavily on starting values. Provide MLE for other
parameters, especially \eqn{\beta_{IE}}, at the same time.
}
\details{
The inverse in \code{fobj} is improved. \eqn{log|M|} is additionally added to
 \code{obj1, obj5}. \code{symmetric=F} is set in \code{eigen}.
}
