% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{incoef_ie}
\alias{incoef_ie}
\title{Data Generation for Inner Envelope Simulation}
\usage{
incoef_ie(SAM = F, u1, u2 = "Hsg4", p, r, n = 400, x = "Hsg4", dia)
}
\arguments{
\item{u1}{Taking value of 1 or 4 or 7, according to the paper.}
}
\value{
\code{list(M=Sig, U=U, beta=beta, Gamma1=Gam1, Gamma12=Gam, Gamma3=Gam3,
 Gamma0=Gam0)} for population version.

 And \code{list(beta=betaOLS, M=M, U=U, Y=Y, X=X, epsilon=eps, Gamma1=Gam1,
 Gamma12=Gam, Gamma3=Gam3, Gamma0=Gam0, coef=coef))} for sample version.
}
\description{
Implement Section 4.1 in the paper of inner envelope (Su, Cook, 2011).
}
\details{
The inputs are set to conform to that of udfenv_coef(), so some of them are
actually useless. And they are preset to be "Hsg4".
}
