% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{init.ienv}
\alias{init.ienv}
\title{Starting values for inner envelope subspace estimation!}
\usage{
init.ienv(M, U, d, p, Gam0.t = NULL, PRINT = F)
}
\arguments{
\item{d}{The dimension of the inner envelope subspace.}

\item{Gam0.t}{True value of Gamma0 as to print the process when \code{PRINT=TRUE}.}

\item{PRINT}{Logical. Whether to print? Will print the Frobenius distance
between each of the starting values candidates and the true value, and the
value of the objective function with \code{det(M)} added.}
}
\value{
The starting values for Gamma0, value of the MLE of inner envelope
 at the chosen starting point and all the 4 candidates:
 \code{list(init=init_min, obj=obj_min, init1=init1, init2=init2, init3=init3,
init4=init4, init5=init5))}.
}
\description{
A fairly good algorithm by choosing from 4 candidates.
}
\details{
When comparing the values of objective functions at the candidate points,
module is used as \code{eigen} always returns complex eigenvalues.

The four candidates are:
\itemize{
  \item Standardized U by eigenvectors of \code{M},
  \item Normal U by eigenvectors of \code{M},
  \item Standardized U by eigenvectors of \code{M+Q_U},
  \item Normal U by eigenvectors of \code{M+Q_U}.
}
}
