% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{init.ienv2}
\alias{init.ienv2}
\title{Revised version of \code{\link{init.ienv}}}
\usage{
init.ienv2(M, U, d, p, c, Gam0.t = NULL, PRINT = F)
}
\arguments{
\item{c}{Scale added to \eqn{Q_U}.}
}
\value{
\item{init}{\eqn{\Gamma_0}.}
\item{M_cQU}{\eqn{M+c*Q_U}.}
}
\description{
Use \code{M+c*Qu} instead of \code{M+Qu}.
}
\details{
The four candidates are:
\itemize{
  \item Standardized U by eigenvectors of \code{M},
  \item Normal U by eigenvectors of \code{M},
  \item Standardized U by eigenvectors of \code{M+c*Q_U},
  \item Normal U by eigenvectors of \code{M+c*Q_U}.
}
}
