% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{init.ienv3}
\alias{init.ienv3}
\title{OBSOLETE}
\usage{
init.ienv3(M, U, d, p, Gam0.t = NULL, PRINT = F, c)
}
\description{
Added a new candidate, selected by the new idea: compress \eqn{U} to make it
reduce \eqn{M}.
}
\details{
Invocation: \code{\link{stv3}}. And to estimate \eqn{A_0}, only the normal
\eqn{U} in \code{stv3} is used.

Does not work when \eqn{d=p}.
}
