% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{innerPLS}
\alias{innerPLS}
\title{The Inner PLS Algorithm}
\usage{
innerPLS(X, Y, d)
}
\arguments{
\item{d}{Dimension of the inner predictor envelope subspace. It should hold
that \eqn{0 \leq d \leq r}.}
}
\value{
\item{Gamma1}
\item{Gamma0}
\item{betaiPLS}
\item{alphaiPLS}
\item{Wu}{The original \eqn{W_u} matrix from \code{SIMPLS_gamma}.}
\item{paras}{The results from \code{\link{IPE}}. Includes some other parameters not so often used.}
}
\description{
Invokes \code{\link{SIMPLS_gamma}} and \code{\link{IPE}}.
}
