% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{IPE}
\alias{IPE}
\title{Inner Predictor Envelope Estimation}
\usage{
IPE(
  X,
  Y,
  d,
  initial = NULL,
  tol = .Machine$double.eps^0.75,
  NonGrass = F,
  ftol.nonG = 0.1,
  PRINT = F
)
}
\arguments{
\item{d}{Dimension of the inner predictor envelope subspace. It should hold
that \eqn{0\le d \le r}.}

\item{initial}{Starting values for \eqn{\Gamma_1}. If set, will directly go
to the Non-Grassmann stage. When \code{d=0}, this will not be used, so no
specifying is fine.}

\item{NonGrass}{Whether to implement the optimization?}

\item{ftol.nonG}{Used in \code{ftol} in \code{\link{obj.IPE}}.}
}
\value{
\item{Gamma1}{Will be \code{NULL} if \eqn{d=0}.}
\item{Gamma0}{Will be an orthogonal matrix if \eqn{d=0}.}
\item{betaIPE}{The IPE estimator of \eqn{\beta}.}
\item{MLE_otherpars}{Estimator for the other parameters. See details in
\code{\link{mle_after_Gamma1.IPE}}.}
\item{alphaIPE}{The intercept.}
\item{loglik}
}
\description{
Overall estimation. Either mere using the starting values or further optimizing
by non-Grassmann technique are both good.
}
