% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alternating Algorithm.R
\name{Lu_ca}
\alias{Lu_ca}
\title{Compute value of the final objective function over C_A with C_{B_1} fixed.}
\usage{
Lu_ca(u1, u2, r, M, invMU, ca, ce, cb1, GEidx.g1, GEidx.g12, GEidx.cb1, obj1)
}
\description{
For other information, see \code{\link{Lu_cb1}}. These two functions are paired,
having the same inputs and outputs.
}
