% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alternating Algorithm.R
\name{Lu_ca.rm1}
\alias{Lu_ca.rm1}
\title{Compute value of the final objective function over C_A with C_{B_1} fixed.}
\usage{
Lu_ca.rm1(
  u1,
  u2,
  r,
  M,
  invMU,
  ca,
  ce,
  cb1,
  GEidx.g1,
  GEidx.g12,
  GEidx.cb1,
  obj1
)
}
\description{
Paired with \code{\link{Lu_cb1.rm1}}.
}
