% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alternating Algorithm.R
\name{Lu_cb1}
\alias{Lu_cb1}
\title{Compute value of the final objective function over C_{B_1} with C_A fixed.}
\usage{
Lu_cb1(u1, u2, r, M, invMU, ca, ce, cb1, GEidx.g1, GEidx.g12, GEidx.cb1, obj1)
}
\arguments{
\item{u1, u2}{u1+u2!=r. While u1+u2<r-1 is also allowed.}
}
\value{
C_A, C_{B_1}, C_E, Gamma1/2/12 and the value of the minimized objective function.
}
\description{
This function is paired with \code{\link{Lu_ca}}.
}
