% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{mle_after_Gamma1.IPE}
\alias{mle_after_Gamma1.IPE}
\title{MLE of other parameters other than Gamma_1 of IPE.}
\usage{
mle_after_Gamma1.IPE(
  SX,
  SXY,
  SYvX,
  Gam1,
  p = dim(Gam1)[1],
  d = dim(Gam1)[2],
  r
)
}
\arguments{
\item{r}{Must be specified.}
}
\value{
\item{betaIPE}
\item{betaOLS}
\item{SiYvX.IPE}
\item{U}{The \eqn{U} matrix. Semi-orthogonal.}
\item{C.IPE}{Will be \code{NA} if \eqn{d=r}.}
}
\description{
The \eqn{\beta}, OLS and IPE ones, are of the transposed version. That is,
the one used in Cook's notation for predictor envelopes.
}
