% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE Non-Grassmann&others.R
\name{mle_after_Gamma1.IRE}
\alias{mle_after_Gamma1.IRE}
\title{MLE of other parameters other than Gamma_1.}
\usage{
mle_after_Gamma1.IRE(
  M,
  U,
  Gam1,
  betaOLS,
  p,
  u1 = dim(Gam1)[2],
  r = dim(Gam1)[1]
)
}
\arguments{
\item{M, U}{The \code{M} and \code{U} matrix.}

\item{p}{Must be inputted.}

\item{u1}{The dimension of \eqn{S_1} subspace. For historical reason, this is named
as u1, but not \code{d}. Should I change it?}
}
\value{
\item{betaIRE}
\item{Omega0.IRE}
\item{B.IRE}
}
\description{
MLE of other parameters other than Gamma_1.
}
