% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{obj.IPE}
\alias{obj.IPE}
\title{Optimization of the Objective Function of IPE by Non-Grassmann Technique.}
\usage{
obj.IPE(
  SX,
  SXY,
  SYvX,
  d,
  initial,
  ftol = 0.1,
  PRINT = F,
  tol = .Machine$double.eps^0.75
)
}
\arguments{
\item{initial}{Starting values for \strong{\eqn{\Gamma_1}} and must be specified.}

\item{ftol}{Used in \code{abs(obj1-obj5)<ftol*abs(obj1)}.}

\item{PRINT}{Whether to print the value of the objective function in each iteration?}

\item{tol}{Used in \code{\link{Pr}}.}
}
\value{
\code{Gamma1, Gamma0, objfun}.
}
\description{
Estimating inner predictor envelope subspace using non-Grassmann technique, which, on
on the same time, relies heavily on starting values.
}
\details{
May need improvement.
}
