% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminaries.R
\name{Pr}
\alias{Pr}
\title{Projection Matrix}
\usage{
Pr(
  mat,
  prodmat = diag(dim(mat)[1]),
  rank = NULL,
  tol = .Machine$double.eps^0.75
)
}
\arguments{
\item{prodmat}{The projection is relative to what inner product? The standard
inner product, the default.}

\item{rank}{Rank of the matrix. Used in \code{\link{mpinv}}. If null, will be
selected in \code{\link{mpinv}}.}

\item{tol}{Set it case by case.}
}
\description{
Projection Matrix
}
