% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{prederr.cv.sim}
\alias{prederr.cv.sim}
\title{Cross-validation prediction error for inner PLS, IPE, Predictor Envelope and SIMPLS}
\usage{
prederr.cv.sim(
  X,
  Y,
  d.inp,
  d.IPE,
  u.predenv,
  u.PLS,
  kfold = 3,
  nperm = 50,
  NonGrass = F,
  PRINT = F
)
}
\arguments{
\item{X, Y}{The data matrix.}
}
\value{
\item{pred_error.inp}
\item{pred_error.IPE}
\item{pred_error.predenv}
\item{pred_error.PLS}
\item{pred_error.predenv2}
\item{pred_error.PLS2}
}
\description{
Compute the prediction error for the envelope estimator using \code{kfold}
cross-validation.
}
