% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{rd_alt_alg.test}
\alias{rd_alt_alg.test}
\title{Alternating Algorithm Simulation}
\usage{
rd_alt_alg.test(
  type = "pop",
  u1 = 10,
  u2 = 10,
  p = 15,
  r = 30,
  n = 500,
  x = 400,
  dia = 1.3^(30:1),
  DTB = F,
  PRINT = F
)
}
\arguments{
\item{type}{This input must exist so it is always possible to check if the codes are right by running them on population version.}

\item{u1, u2}{u1+u2!=r is the only constraint.}

\item{DTB}{When disturbance is used, envMU(initial=starting values.disturbed). While when DTB=F, envMU(initial=Gam1.sv) when estimating Gam1, where Gam1.sv comes from starval().}
}
\description{
Print the result of sv,clc and rd estimators and the estimation of beta by alternating algorithm.
}
