% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation of Simulation.R
\name{results_save}
\alias{results_save}
\title{Write Results to Files}
\usage{
results_save(
  result_Simulation,
  fdnm,
  BOXPLOT = T,
  estinm = c("OLS", result_Simulation$estmrs),
  dat = result_Simulation$dat,
  work_dir0 = "Y:/Research/Data Generated and Results/IPE",
  ylim_boxplot = c(0, max(result_Simulation$dis)),
  ylab_boxplot = "Bias",
  ...
)
}
\arguments{
\item{fdnm}{The name of the folder that the data is read from.}

\item{BOXPLOT}{Draw a boxplot of \code{result_simulation$dis}? Use
\code{\link[graphics]{boxplot}}. The arguments are
\code{outline=F, ylim=c(0,max(result_Simulation$dis)), ylab="Bias"}.}

\item{estinm}{The names of the estimators used.}

\item{work_dir0}{The name of the father working directory of \code{fdnm}.}

\item{...}{Optional arguments to \code{\link[graphics]{boxplot}}. For example,
\code{names=c(expression(beta[OLS]), expression(beta[iPLS]), expression(beta[IPE], expression(beta[Predenv])))}.}

\item{result_simulation}{The output from \code{\link{Run_Simulation}}.}
}
\description{
Creates or overwrites a folder named of the type of distribution used.
}
