% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminaries.R
\name{rsomat}
\alias{rsomat}
\title{Generate a random semi-orthogonal matrix.}
\usage{
rsomat(u, r)
}
\arguments{
\item{r, u}{\code{1<=u<=r}.}
}
\value{
A matrix, which is the desired one.
}
\description{
The matrix is generated by standardizing a matrix of independent U(0,1)
variates.
}
