% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation of Simulation.R
\name{Run_Simulation}
\alias{Run_Simulation}
\title{Run the Simulation of inner PLS, IPE, Predictor Envelopes and SIMPLS}
\usage{
Run_Simulation(
  dat,
  rep_num = 5,
  estmrs = c("iPLS", "IPE", "PredEnv", "SIMPLS", "LASSO", "Ridge"),
  SELECT_DIM = F,
  u.PredEnv = NULL,
  u.SIMPLS = NULL,
  dim.iPLS_IPE = c("bootstrap", "CV", "BIC"),
  dim.PredEnv = c("bootstrap", "CV", "BIC"),
  dim.SIMPLS = c("bootstrap", "CV"),
  err_type = c("Normal", "Uniform"),
  coef_normal = 1,
  upper.uniform,
  B = 50,
  kfold = 3,
  n = 1000,
  NonGrass = F,
  ftol.nonG = 0.5,
  lambda_seq = seq(from = 0, to = 20, by = 0.2)
)
}
\arguments{
\item{dat}{A list containing the parameters of a inner predictor envelope.}

\item{rep_num}{Replication times of the simulation.}

\item{estmrs}{The estimators to use. \code{PredEnv} depends on package \code{Renvlp},
and \code{LASSO,Ridge} depend on package \code{glmnet}.}

\item{SELECT_DIM}{Whether to perform dimension selection. If \code{FALSE}, true
value is used.}

\item{u.PredEnv, u.SIMPLS}{Specified dimension to be used for predictor envelopes
and SIMPLS. Valid only when \code{SELECT_DIM=FALSE}.}

\item{dim.iPLS_IPE}{The type of dimension selection method to use for inner PLS and SIMPLS.
If set to \code{CV}, both estimators must be specified in \code{estmrs}.
If set to \code{BIC}, iPLS is proceeded by bootstrap.}

\item{dim.PredEnv}{Similar to \code{dim.SIMPLS}.}

\item{dim.SIMPLS}{The type of dimension selection method to use for SIMPLS.
If "bootstrap", bootstrap is used for \code{SELECT_DIM=TRUE} and
true value otherwise. If \code{estmrs} does not contain \code{"SIMPLS"}, this
argument does not matter.}

\item{err_type}{The error type to use, as listed now. If more than one, the
first will be used.}

\item{coef_normal}{Coefficient to time the SiYvX matrix.}

\item{upper.uniform}{Will generate \code{U(0, upper.uniform)} errors for uniform
errors.}

\item{B}{The bootstrap replication used in dimension selection.}

\item{kfold}{The cross-validation folds used in dimension selection, for SIMPLS.}

\item{NonGrass}{Whether to further implement optimization after obtaining the
starting values of IPE.}

\item{lambda_seq}{Provided to \code{\link[glmnet]{cv.glmnet}} as \code{lambda=lambda_seq}
for LASSO and Ridge.}
}
\value{
A list. Simulation results and arguments used. Important ones are:
\item{dis}{Distance between estimators(including OLS) of beta and the real
one in each replication.}
\item{dimen}{Dimension selected for estimators in each replication}
\item{mean_dis}
\item{mean_dimen}
\item{MSE}{The mean squared error of the estimators.}
\item{time_used}
\item{dimen_dis.iPLS}{The bootstrap vector correlations of each dimension in each
 replication. If CV is used, this does not appear. If dimension selection is
 performed, \code{NA} is returned.}
\item{mean_dimen_dis.iPLS}
}
\description{
Run the Simulation of inner PLS, IPE, Predictor Envelopes and SIMPLS.
Coded vectorizingly. Should be quick.
}
\details{
Prints "-" each time \code{\link{d.boot.sim}} completes in a replication.
}
