% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implementation of Simulation.R
\name{Run_Simulation0}
\alias{Run_Simulation0}
\title{Run the Simulation of inner PLS, IPE, Predictor Envelopes and SIMPLS}
\usage{
Run_Simulation0(
  dat,
  rep_num = 5,
  estmrs = c("iPLS", "IPE", "PredEnv", "SIMPLS"),
  SELECT_DIM = T,
  err_type = c("normal", "uniform"),
  coef_normal = 1,
  upper.uniform,
  B = 50,
  n = 1000,
  NonGrass = F,
  ftol.nonG = 0.5
)
}
\arguments{
\item{dat}{A list containing the parameters of a inner predictor envelope.}

\item{rep_num}{Replication times of the simulation.}

\item{estmrs}{The estimators to use.}

\item{SELECT_DIM}{Whether to select dimensions of the estimators. The dimension
selection is done by \code{\link{d.boot.sim}}.}

\item{err_type}{The error type to use, as listed now. If more than one, the
first will be used.}

\item{coef_normal}{Coefficient to time the SiYvX matrix.}

\item{upper.uniform}{Will generate \code{U(0, upper.uniform)} errors for uniform
errors.}

\item{B}{The bootstrap replication used in dimension selection.}

\item{NonGrass}{Whether to further implement optimization after obtaining the
starting values of IPE.}
}
\value{
A list. Objects with \code{.iPLS} have counterparts with \code{.IPE, .predenv, .PLS}.
\item{dis}{Distance between estimators of beta and the real one in each replication.}
\item{dimen}{Dimension selected for estimators in each replication}
\item{mean_dis}
\item{mean_dimen}
\item{MSE}{The mean squared error of the estimators.}
\item{dimen.iPLS}{The bootstrap vector correlations of each dimension in each
 replication. Will be \code{NA} if \code{SELECT_DIM=F}.}
\item{mean_dimen.iPLS}
\item{time_used}
}
\description{
OBSOLETE because too slow due to endless "if/else" loop.
}
\details{
Four possible estimators: inner PLS, IPE, Predictor Envelopes and SIMPLS.
}
