% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/For Simulations.R
\name{scenarios_nonGrassmann}
\alias{scenarios_nonGrassmann}
\title{CLC Envelope Simulation}
\usage{
scenarios_nonGrassmann(case)
}
\value{
\code{M, U, betaOLS, u, r, Gamma1=Gamma1, beta=beta, X, Y,
  epsilon}.
}
\description{
Implement Section 4.1 in the paper of non-Grassmann algorithm (A Note on Fast
Envelope Estimation, Cook, Forzani and Su, 2016).
}
