% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other Algorithms.R
\name{SIMPLS_env}
\alias{SIMPLS_env}
\title{SIMPLS for envelope regression estimation}
\usage{
SIMPLS_env(X, Y, u)
}
\value{
\item{beta.PLS}
\item{Gamma.PLS}
\item{mu.PLS}
}
\description{
Deals with sample estimation. Invokes \code{\link{SIMPLS_gamma}} to calculate \eqn{\Gamma}.
}
