% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other Algorithms.R
\name{SIMPLS_gamma}
\alias{SIMPLS_gamma}
\title{SIMPLS for envelope estimation}
\usage{
SIMPLS_gamma(M, U, u)
}
\value{
\item{Gammahat}{Estimator for the envelope subspace.}
\item{Wu}{The original \eqn{W_u} matrix.}
}
\description{
Specifically, the algorithm in Section 6.5.1, Cook's Book.
}
