% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other Algorithms.R
\name{SIMPLS_gamma2}
\alias{SIMPLS_gamma2}
\title{Improvement of \code{\link{SIMPLS_gamma}}.}
\usage{
SIMPLS_gamma2(M, U, u)
}
\description{
Use eigenvectors in calculating \eqn{Q_{Ecal}}. But it seems this does not make
a different at all.
}
