% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{starval}
\alias{starval}
\title{Get starting values for RD Envelope.}
\usage{
starval(M, U, beta, invMU = NULL, u1, u2, r)
}
\arguments{
\item{u1, u2, r}{u1+u2=r-1 is allowed!}
}
\value{
Gamma1, Gamma2, Gamma0. Now Gamma2 is strictly inside Gamma12, while two versions of Gamma1 are provided.
}
\description{
Get the starting values for Gamma1 and Gamma12.
}
\details{
Invocation: \code{\link{stv}}, \code{\link[Renvlp]{ginv}}.

\eqn{Q_\beta} is used when estimating Gamma1. And \code{tol=.Machine$double.eps^0.8} is used in \code{ginv} during this computation.
}
