% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{starval2}
\alias{starval2}
\title{Revised version of \code{\link{starval}}.}
\usage{
starval2(M, U, beta, invMU = NULL, u1, u2, r, Gam1.t, Gam12.t)
}
\description{
The only difference is that this function invocates \code{\link{stv2}} instead of \code{\link{stv}}, with corresponding changes in inputs.
Things printed can be seen in \code{\link{stv2}}.
}
\details{
Gam1 from Pu is used.
}
