% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{stv}
\alias{stv}
\title{Generic starting values computation function.}
\usage{
stv(M, U, u, Gam.t = NULL)
}
\arguments{
\item{U}{U is not of full rank, or it follows that u2=0.}

\item{u}{u<r.}

\item{Gam.t}{A redundant input, to make inputs conform to that of \code{\link{stv2}}.}
}
\value{
The starting values and value of the minimized objective function:
\code{list(init=init1, obj=obj1)}.
}
\description{
This function is completely copied from \code{\link[Renvlp]{envMU}} only with
the names of the variable revised.
}
