% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{stv3}
\alias{stv3}
\title{Improved version of \code{\link{stv}}}
\usage{
stv3(M, U, u, USE = 1:4, Gam.t, PRINT = F)
}
\arguments{
\item{USE}{A vector, each of whose element taking value from 1 to 4. Specify
which candidates to use.}
}
\description{
Which of four candidates to use can be specified.
}
\details{
The 4 potential starting values are:
\itemize{
  \item Normal U by eigenvectors of \code{M+U},
  \item Standardized U by eigenvectors of \code{M+U},
  \item Normal U by eigenvectors of \code{M},
  \item Standardized U by eigenvectors of \code{M}.
}
}
