% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starting Values.R
\name{sv.IPE}
\alias{sv.IPE}
\title{Starting Values for Inner Predictor Envelopes.}
\usage{
sv.IPE(
  SX,
  SXY,
  SYvX,
  d,
  r,
  c,
  Gam1.t = NULL,
  PRINT = F,
  tol = .Machine$double.eps^0.75
)
}
\arguments{
\item{SYvX}{\eqn{S_{Y|X}}.}

\item{c}{Temporarily obsolete.}

\item{tol}{Used in \code{\link{Pr}}.}
}
\value{
\item{init}{\eqn{\Gamma_1}.}
}
\description{
Only two types of starting values: both using eigenvectors from SX.
Mainly copied from \code{\link{init.ienv2}}.
}
\details{
The two candidates are:
\itemize{
  \item Standardized U by eigenvectors of \code{M},
  \item Normal U by eigenvectors of \code{M}.
}
}
