% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{u.cv.PredEnv}
\alias{u.cv.PredEnv}
\title{Dimension Selection of Predictor Envelope by Cross Validation}
\usage{
u.cv.PredEnv(X, Y, kfold = 3, trace = 0)
}
\value{
\item{u.cv.PredEnv}
\item{prederr.PredEnv}
}
\description{
Based on \code{\link{u.cv.sim}}.
}
