% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{u.cv.sim}
\alias{u.cv.sim}
\title{Dimension selection of inner PLS, IPE by cross-validation.}
\usage{
u.cv.sim(X, Y, kfold = 3, trace = 0, NonGrass = F)
}
\arguments{
\item{kfold}{The folds to use in CV.}

\item{trace}{If 1, a "-" is printed after each fold is completed.
If -1, a "-" is printed after the whole function ends.}
}
\value{
\item{d.cv.iPLS}{The selected dimension.}
\item{prederr.iPLS}{Prediction errors of each dimension.}
\item{d.cv.IPE}
\item{prederr.IPE}
}
\description{
The range is from \eqn{0} to \eqn{r}.
}
