% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation for IPE.R
\name{u.cv.SIMPLS}
\alias{u.cv.SIMPLS}
\title{Dimension selection of SIMPLS by cross-validation.}
\usage{
u.cv.SIMPLS(X, Y, kfold = 3, trace = 0)
}
\arguments{
\item{trace}{If 1, a "-" is printed after each fold is completed.
If -1, a "-" is printed after the whole function ends.}
}
\value{
\item{u.cv.SIMPLS}
\item{pred_err}
}
\description{
Dimension selection of SIMPLS by cross-validation.
}
