% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminaries.R
\name{vec_cor}
\alias{vec_cor}
\title{Vector Correlation Coefficient}
\usage{
vec_cor(A, B)
}
\arguments{
\item{A, B}{Semi-orthogonal matrices of the same dimension.}
}
\description{
Vector Correlation Coefficient
}
